% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_lm.R
\name{scatter_lm}
\alias{scatter_lm}
\title{Create a scatter plot and fitted linear model}
\usage{
scatter_lm(x, y, df)
}
\arguments{
\item{x}{Class name of x-axis of dataset.}

\item{y}{Class name of y-axis of dataset.}

\item{df}{Name of dataset.}
}
\value{
A scatter plot containing a fitted linear model.
}
\description{
Create a scatter plot and fitted linear model
}
